Nickreal1

+++ src/Model/Indexer.java
@@ -704,11 +704,11 @@ public class Indexer {
     public boolean reset() {
         File fileTo = new File(pathTo);
         File[] filesInPathTo = fileTo.listFiles();
-        boolean succDelete = false;
         for (File file : filesInPathTo) {
-            if (file.isDirectory() && file.listFiles().length > 0)
+            if (file.isDirectory() && (file.getName().equals("WithoutStemming") || file.getName().equals("WithStemming")) && file.listFiles().length > 0)
                 recDelete(file.listFiles());
-            file.delete();
+            if(file.isDirectory() && (file.getName().equals("WithoutStemming") || file.getName().equals("WithStemming")))
+               file.delete();
         }
         System.gc();
         return true;
